refactor(config): one file, one helper, and an environment name on every number
Kai asked for all of the configuration in config.go and all of it through the
same overridable helper. This is both halves.

`tuning.go` is gone. Its numbers live in `config.go` beside what the deployment
supplies, and every one of them is declared on a line that binds three things:
the variable the package reads, the environment name that sets it, and the
default it holds without one. There is no longer a tier of numbers a deployment
cannot reach.

The helper is generic over the two shapes a number takes here, a count and a
duration, so a count is as settable as a timeout and neither has a parser of
its own. The second hand-maintained table of pointers is gone with it: adding a
number to the table is now the only way to add one, which is what keeps the
list complete.

One behaviour is now uniform that was not. `REQUEST_TIMEOUT`, `QUEUE_TIMEOUT`,
and `SHUTDOWN_GRACE` were parsed a second time to fill a Config field, and that
second reader refused a bad value and failed the load while the table applied
one silently. Config now takes what the knob pass produced, so a bad value
keeps the default everywhere. Silence there would read as a working override,
so applied and rejected names are reported on the capabilities log line.

The list is generated. `ward exec knobs` writes docs/sirens-echo-knobs.md from
the table and `knobs-check` fails when it is stale, so no inventory is
maintained by hand.

Three numbers stay derived, because a value defined as an expression of another
is set by setting its input. Two algorithm floors stay where they are used:
they decide what counts as a match rather than how much of one to allow.

+++ internal/community/readiness.go
@@ -94,7 +94,7 @@ func (a *Agent) checkRouteReadiness(ctx context.Context) readinessOutcome {
 		return readinessDependencyError
 	}
 	defer response.Body.Close()
-	raw, err := io.ReadAll(io.LimitReader(response.Body, maxReadinessBody+1))
+	raw, err := io.ReadAll(io.LimitReader(response.Body, int64(maxReadinessBody)+1))
 	if err != nil || len(raw) > maxReadinessBody {
 		return readinessInvalidResponse
 	}
